refactor: align api package with main
