+++ Dockerfile.api
@@ -1,9 +1,9 @@
 FROM python:3.11-slim
 
 WORKDIR /app
-COPY requirements.txt .
+COPY api/requirements.txt ./requirements.txt
 RUN pip install --no-cache-dir -r requirements.txt
-COPY trailguard_api /app/trailguard_api
+COPY api /app/api
 COPY openapi.yaml /app/openapi.yaml
 COPY migrations /app/migrations
-CMD ["uvicorn", "trailguard_api.main:app", "--host", "0.0.0.0", "--port", "8000"]
+CMD ["uvicorn", "api.main:app", "--host", "0.0.0.0", "--port", "8000"]
